pyproject.toml
[tool.poetry]
name = "coursework"
version = "0.1.0"
description = ""
authors = ["stanislav <stasfed16@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.1.1"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
mypy = "^1.11.2"
black = "^24.8.0"
pytest = "^8.3.3"
requests = "^2.32.3"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
)
'''


[tool.isort]
line-lenght = 119
